kautools: Add options to \listofpapers command.
This adds options to \listofpapers that controls whether to add a TOC
entry for the list of papers, and whether to print the participation
information from the defined papers. The options are called
'tocline' and 'printparticipation' and both default to yes.

Turning off the former can be useful if, for instance, the list of
papers comes before the table of contents. Turning off the latter can be
useful if the participation information is repetitive, in which case
replacing the per-paper listing with a self-written block of text that
describes all of the papers can make sense.